rename ports component
