+++ pages/c/_cluster/workloads/_namespace/_id.vue
@@ -57,6 +57,6 @@ export default {
     </header>
     <ResourceYaml v-if="asYaml" :obj="obj" :value="yaml" :done-route="doneRoute" :for-create="true" />
 
-    <Workload v-else :value="obj" :mode="mode" />
+    <Workload v-else :value="obj" :mode="mode" :namespace-suffix-on-create="true" />
   </div>
 </template>
